Things (#39)
* Packages code better

* This better work

* Uuuuuuuuugh

* more

* Documentation
