+++ __init__.py
@@ -1,8 +1,1 @@
-from Cayley.abstractnetwork import *
-from Cayley.cayleytree import *
-from Cayley.graph import *
-from Cayley.lattice import *
-from Cayley.montecarlo import *
-
-
-
+from Cayley.classes import *
