+++ classes/abstractnetwork.py
